+++ dpws/src/main/java/org/somda/sdc/dpws/client/helper/HostingServiceResolver.java
@@ -18,7 +18,7 @@ import org.somda.sdc.dpws.TransportBinding;
 import org.somda.sdc.dpws.client.DiscoveredDevice;
 import org.somda.sdc.dpws.factory.CommunicationLogFactory;
 import org.somda.sdc.dpws.factory.TransportBindingFactory;
-import org.somda.sdc.dpws.guice.NetworkJobThreadPool;
+import org.somda.sdc.dpws.guice.ResolverThreadPool;
 import org.somda.sdc.dpws.http.HttpUriBuilder;
 import org.somda.sdc.dpws.model.HostServiceType;
 import org.somda.sdc.dpws.model.HostedServiceType;
@@ -65,7 +65,7 @@ import java.util.concurrent.TimeoutException;
 public class HostingServiceResolver {
     private static final Logger LOG = LogManager.getLogger(HostingServiceResolver.class);
 
-    private final ExecutorWrapperService<ListeningExecutorService> networkJobExecutor;
+    private final ExecutorWrapperService<ListeningExecutorService> resolveExecutor;
     private final LocalAddressResolver localAddressResolver;
     private final TransportBindingFactory transportBindingFactory;
     private final RequestResponseClientFactory requestResponseClientFactory;
@@ -84,7 +84,7 @@ public class HostingServiceResolver {
 
     @Inject
     HostingServiceResolver(@Named(DpwsConfig.MAX_WAIT_FOR_FUTURES) Duration maxWaitForFutures,
-                           @NetworkJobThreadPool ExecutorWrapperService<ListeningExecutorService> networkJobExecutor,
+                           @ResolverThreadPool ExecutorWrapperService<ListeningExecutorService> resolveExecutor,
                            LocalAddressResolver localAddressResolver,
                            TransportBindingFactory transportBindingFactory,
                            RequestResponseClientFactory requestResponseClientFactory,
@@ -102,7 +102,7 @@ public class HostingServiceResolver {
         this.communicationLogFactory = communicationLogFactory;
         this.instanceLogger = InstanceLogger.wrapLogger(LOG, frameworkIdentifier);
         this.maxWaitForFutures = maxWaitForFutures;
-        this.networkJobExecutor = networkJobExecutor;
+        this.resolveExecutor = resolveExecutor;
         this.localAddressResolver = localAddressResolver;
         this.transportBindingFactory = transportBindingFactory;
         this.requestResponseClientFactory = requestResponseClientFactory;
@@ -131,7 +131,7 @@ public class HostingServiceResolver {
      * @return Future with resolved hosting service and hosted service information.
      */
     public ListenableFuture<HostingServiceProxy> resolveHostingService(DiscoveredDevice discoveredDevice) {
-        return networkJobExecutor.get().submit(() -> {
+        return resolveExecutor.get().submit(() -> {
             if (discoveredDevice.getXAddrs().isEmpty()) {
                 throw new IllegalArgumentException("Given device proxy has no XAddrs. Connection aborted.");
             }