+++ dpws/src/main/java/org/somda/sdc/dpws/guice/DefaultDpwsModule.java
@@ -111,6 +111,7 @@ public class DefaultDpwsModule extends AbstractModule {
     private ExecutorWrapperService<ScheduledExecutorService> appDelayExecutor;
     private ExecutorWrapperService<ListeningExecutorService> networkJobThreadPoolExecutor;
     private ExecutorWrapperService<ListeningExecutorService> wsDiscoveryExecutor;
+    private ExecutorWrapperService<ListeningExecutorService> resolveExecutor;
 
     /**
      * Constructor.
@@ -119,6 +120,7 @@ public class DefaultDpwsModule extends AbstractModule {
         appDelayExecutor = null;
         networkJobThreadPoolExecutor = null;
         wsDiscoveryExecutor = null;
+        resolveExecutor = null;
     }
 
     @Override
@@ -352,5 +354,25 @@ public class DefaultDpwsModule extends AbstractModule {
         }
 
         return wsDiscoveryExecutor;
+    }
+
+    @Provides
+    @ResolverThreadPool
+    ExecutorWrapperService<ListeningExecutorService> getResolverThreadPool(@Named(CommonConfig.INSTANCE_IDENTIFIER)
+                                                                                     String frameworkIdentifier) {
+        if (resolveExecutor == null) {
+            Callable<ListeningExecutorService> executor = () -> MoreExecutors.listeningDecorator(Executors
+                    .newFixedThreadPool(
+                            10,
+                            new ThreadFactoryBuilder()
+                                    .setNameFormat("ResolverThreadPool-thread-%d")
+                                    .setDaemon(true)
+                                    .build()
+                    ));
+            resolveExecutor = new ExecutorWrapperService<>(executor, "ResolverThreadPool",
+                    frameworkIdentifier);
+        }
+
+        return resolveExecutor;
     }
 }